+++ packages/extensions/src/extensions.ts
@@ -10,6 +10,7 @@ import { MaterialsSpecular } from './khr-materials-specular';
 import { MaterialsTransmission } from './khr-materials-transmission';
 import { MaterialsUnlit } from './khr-materials-unlit';
 import { MaterialsVariants } from './khr-materials-variants';
+import { MaterialsVolume } from './khr-materials-volume';
 import { MeshQuantization } from './khr-mesh-quantization';
 import { TextureBasisu } from './khr-texture-basisu';
 import { TextureTransform } from './khr-texture-transform';
@@ -25,6 +26,7 @@ export const KHRONOS_EXTENSIONS = [
 	MaterialsTransmission,
 	MaterialsUnlit,
 	MaterialsVariants,
+	MaterialsVolume,
 	MeshQuantization,
 	TextureBasisu,
 	TextureTransform,
@@ -48,6 +50,7 @@ export * from './khr-materials-pbr-specular-glossiness';
 export * from './khr-materials-transmission';
 export * from './khr-materials-unlit';
 export * from './khr-materials-variants';
+export * from './khr-materials-volume';
 export * from './khr-mesh-quantization';
 export * from './khr-texture-basisu';
 export * from './khr-texture-transform';