+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 **Features:**
 
 - Add 'resample' transform.
+- Add `KHR_materials_volume` extension.
 - Add `EXT_mesh_gpu_instancing` extension.
 
 **Breaking changes:**